+++ setup.cfg
@@ -29,7 +29,7 @@ install_requires =
     astropy-healpix
     cached-property
     deprecation
-    hera-cli-utils
+    hera-cli-utils>=0.1.0
     numpy>=1.18
     pyuvdata
     pyuvsim>=1.2.5
